backend/requirements.txt
Django>=4.2,<5.0
wagtail>=5.2
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8
psycopg2-binary>=2.9